Notification system implemented

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -71,19 +71,35 @@ export async function POST(
       return Response.json({error: "Post not found"}, {status: 404});
     }
 
-    await prisma.like.upsert({
-      where: {
-        userId_postId: {
+
+    /* Transaction is all or nothing wrapper */
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: {
+          userId_postId: {
+            userId: loggedInUser.id,
+            postId
+          }
+        },
+        create: {
           userId: loggedInUser.id,
           postId
-        }
-      },
-      create: {
-        userId: loggedInUser.id,
-        postId
-      },
-      update: {} // No need to update anything
-    })
+        },
+        update: {} // No need to update anything
+      }),
+      ...(loggedInUser.id !== post.userId // Don't notify the user if they liked their own post
+        ? [
+          prisma.notification.create({
+            data: {
+              issuerId: loggedInUser.id,
+              recipientId: post.userId,
+              postId,
+              type: "LIKE"
+           }
+          })
+          ]
+        : []),
+    ])
 
     return new Response();
 
@@ -104,12 +120,33 @@ export async function DELETE(
       return Response.json({error: "Unauthorized"}, {status: 401});
     }
 
-    await prisma.like.deleteMany({
-      where: {
-        userId: loggedInUser.id,
-        postId
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
       }
     });
+
+    if (!post) {
+      return Response.json({error: "Post not found"}, {status: 404});
+    }
+
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: loggedInUser.id,
+          postId
+        }
+      }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          postId,
+          type: "LIKE"
+        }
+      })
+    ])
 
     return new Response();
